PR feedback
Lint

Arrow function

Move some functions to utils

Usage of arrow functions

Revert package downgrade

Remove ai from custom paths tsconfig

Move normalize line endings and test

node: imports

Fix paths, remove vitest imports
