+++ package.json
@@ -43,7 +43,7 @@
     "@opentelemetry/exporter-trace-otlp-http": "^0.203.0",
     "@opentelemetry/sdk-node": "^0.203.0",
     "@opentelemetry/sdk-trace-base": "^2.0.1",
-    "ai": "^5.0.10",
+    "ai": "^5.0.15",
     "braintrust": "^0.2.1",
     "cors": "^2.8.5",
     "dotenv-flow": "^4.1.0",
