+++ src/mastra/workflows/logCoreAnalyzerWorkflow.ts
@@ -8,7 +8,6 @@ import {
   loadFromFile,
   loadFromUrl,
   loadFromText,
-  normalizeText,
   type LoadResult,
 } from './logCoreAnalyzer/dataLoader';
 import {
@@ -21,6 +20,7 @@ import {
   USER_CONCISE_SUMMARY_PROMPT,
   SINGLE_PASS_PROMPT,
 } from './logCoreAnalyzer/prompts';
+import { normalizeLineEndings } from './logCoreAnalyzer/utils';
 
 // We define here the core workflow for log file analysis. It gives the Parsley Agent the capability to read and understand text files, of any kind and format.
 // Depending on the file size, we either return a summary in a single LLM call, or perform a more complex iterative refinement, combining the usage of cheap and more expensive models.
@@ -82,7 +82,7 @@ const loadDataStep = createStep({
     }
 
     // Normalize the text
-    const normalizedText = normalizeText(result.text);
+    const normalizedText = normalizeLineEndings(result.text);
 
     logger.info('Data loaded successfully', {
       source: result.metadata.source,
